+++ Sources/Example_11/main.swift
@@ -12,47 +12,47 @@ public class Example_11 {
 
         let page = Page(pdf, Letter.PORTRAIT)
 
-        var code = Barcode(Barcode.CODE_128, "Hellö, World!")
+        var code = try Barcode(Barcode.CODE_128, "Hellö, World!")
         code.setLocation(170.0, 70.0)
         code.setModuleLength(0.75)
         code.setFont(f1)
         code.drawOn(page)
 
-        code = Barcode(Barcode.CODE_128, "G86513JVW0C")
+        code = try Barcode(Barcode.CODE_128, "G86513JVW0C")
         code.setLocation(170.0, 170.0)
         code.setModuleLength(0.75)
         code.setDirection(Direction.TOP_TO_BOTTOM)
         code.setFont(f1)
         code.drawOn(page)
 
-        code = Barcode(Barcode.CODE_39, "WIKIPEDIA")
+        code = try Barcode(Barcode.CODE_39, "WIKIPEDIA")
         code.setLocation(270.0, 370.0)
         code.setModuleLength(0.75)
         code.setFont(f1)
         code.drawOn(page)
 
-        code = Barcode(Barcode.CODE_39, "CODE39")
+        code = try Barcode(Barcode.CODE_39, "CODE39")
         code.setLocation(400.0, 70.0)
         code.setModuleLength(0.75)
         code.setDirection(Direction.TOP_TO_BOTTOM)
         code.setFont(f1)
         code.drawOn(page)
 
-        code = Barcode(Barcode.CODE_39, "CODE39")
+        code = try Barcode(Barcode.CODE_39, "CODE39")
         code.setLocation(450.0, 70.0)
         code.setModuleLength(0.75)
         code.setDirection(Direction.BOTTOM_TO_TOP)
         code.setFont(f1)
         code.drawOn(page)
 
-        code = Barcode(Barcode.UPC_A, "51234567890") // UPC-A without the check digit which we calculate!!
+        code = try Barcode(Barcode.UPC_A, "51234567890") // UPC-A without the check digit which we calculate!!
         code.setLocation(450.0, 250.0)
         code.setModuleLength(1.0)
         code.setDirection(Direction.BOTTOM_TO_TOP)
         code.setFont(f1)
         code.drawOn(page)
 
-        code = Barcode(Barcode.EAN_13, "051234567890") // EAN-13 without the check digit which we calculate!!
+        code = try Barcode(Barcode.EAN_13, "051234567890") // EAN-13 without the check digit which we calculate!!
         code.setLocation(450.0, 450.0)
         code.setModuleLength(1.0)
         code.setDirection(Direction.BOTTOM_TO_TOP)